content/canvas/test/test_2d.path.isPointInPath.basic.1.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.isPointInPath.basic.1</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.rect(0, 0, 20, 20);
ok(ctx.isPointInPath(10, 10) === true, "ctx.isPointInPath(10, 10) === true");

if (!_deferred) SimpleTest.finish();
});
</script>
